调整购物车

+++ work-shop/work-shop-core/src/main/java/im/heart/shop/service/CartItemService.java
@@ -6,6 +6,6 @@ import im.heart.shop.entity.CartItem;
 import java.math.BigInteger;
 
 
-public interface CartItemService extends CommonService<CartItem, BigInteger> {
+public interface CartItemService extends CommonService<CartItem, Long> {
 	public static final String BEAN_NAME = "cartItemService";
 }
